Russian Language small fix
After the test in the game, it looks much more logical and the meaning is not lost.

Note!

The Russian translations of the words "Copy" and "Paste" are longer. Characters are more than twice as long! Because of this the words do not fit into the frame completely. To fix this, I shortened the words.